:hammer: [chore] AnniversaryView 추가 #12
